move Tag events to ...casemodule.events package; explicitly implement Serializable interface.
use as needed logger to avoid serializing it.
