+++ Core/src/org/sleuthkit/autopsy/casemodule/events/BlackBoardArtifactTagAddedEvent.java
@@ -7,7 +7,7 @@
  * Licensed under the Apache License, Version 2.0 (the "License");
  * you may not use this file except in compliance with the License.
  * You may obtain a copy of the License at
- * 
+ *
  *     http://www.apache.org/licenses/LICENSE-2.0
  *
  * Unless required by applicable law or agreed to in writing, software
@@ -16,34 +16,36 @@
  * See the License for the specific language governing permissions and
  * limitations under the License.
  */
-package org.sleuthkit.autopsy.events;
+package org.sleuthkit.autopsy.casemodule.events;
 
+import java.io.Serializable;
 import javax.annotation.concurrent.Immutable;
 import org.sleuthkit.autopsy.casemodule.Case;
-import org.sleuthkit.datamodel.ContentTag;
+import org.sleuthkit.datamodel.BlackboardArtifactTag;
 import org.sleuthkit.datamodel.TskCoreException;
 
 /**
- * An event that is fired when a ContentTag is added.
+ * Event sent when a black board artifact tag is added.
  */
 @Immutable
-public class ContentTagAddedEvent extends TagAddedEvent<ContentTag> {
+public class BlackBoardArtifactTagAddedEvent extends TagAddedEvent<BlackboardArtifactTag> implements Serializable {
 
     private static final long serialVersionUID = 1L;
 
-    public ContentTagAddedEvent(ContentTag newTag) {
-        super(Case.Events.CONTENT_TAG_ADDED.toString(), newTag);
+    public BlackBoardArtifactTagAddedEvent(BlackboardArtifactTag newTag) {
+        super(Case.Events.BLACKBOARD_ARTIFACT_TAG_ADDED.toString(), newTag);
     }
 
     /**
-     * get the ContentTag that was added by its id
+     * get the BlackboardArtifactTag that was added by its id
      *
-     * @return ContentTag that was added
+     * @return BlackboardArtifactTag that was added
      *
      * @throws IllegalStateException
      * @throws TskCoreException
      */
-    ContentTag getTagByID() throws IllegalStateException, TskCoreException {
-        return Case.getCurrentCase().getServices().getTagsManager().getContentTagByTagID(getTagID());
+    @Override
+    BlackboardArtifactTag getTagByID() throws IllegalStateException, TskCoreException {
+        return Case.getCurrentCase().getServices().getTagsManager().getBlackboardArtifactTagByTagID(getTagID());
     }
 }
